requirements.txt
Django==5.1.3
celery==5.4.0
django-celery-beat==2.7.0
djangorestframework==3.15.2
elasticsearch==8.16.0
elasticsearch-dsl==8.16.0
redis==5.2.0
black==24.10.0
flake8==7.1.1